update to latest base

+++ src/EVault/shared/lib/RPow.sol
@@ -20,23 +20,30 @@
 pragma solidity ^0.8.0;
 
 library RPow {
-    function rpow(uint x, uint n, uint base) internal pure returns (uint z) {
+    function rpow(uint256 x, uint256 n, uint256 base) internal pure returns (uint256 z) {
         assembly {
-            switch x case 0 {switch n case 0 {z := base} default {z := 0}}
+            switch x
+            case 0 {
+                switch n
+                case 0 { z := base }
+                default { z := 0 }
+            }
             default {
-                switch mod(n, 2) case 0 { z := base } default { z := x }
-                let half := div(base, 2)  // for rounding.
-                for { n := div(n, 2) } n { n := div(n,2) } {
+                switch mod(n, 2)
+                case 0 { z := base }
+                default { z := x }
+                let half := div(base, 2) // for rounding.
+                for { n := div(n, 2) } n { n := div(n, 2) } {
                     let xx := mul(x, x)
-                    if iszero(eq(div(xx, x), x)) { revert(0,0) }
+                    if iszero(eq(div(xx, x), x)) { revert(0, 0) }
                     let xxRound := add(xx, half)
-                    if lt(xxRound, xx) { revert(0,0) }
+                    if lt(xxRound, xx) { revert(0, 0) }
                     x := div(xxRound, base)
-                    if mod(n,2) {
+                    if mod(n, 2) {
                         let zx := mul(z, x)
-                        if and(iszero(iszero(x)), iszero(eq(div(zx, x), z))) { revert(0,0) }
+                        if and(iszero(iszero(x)), iszero(eq(div(zx, x), z))) { revert(0, 0) }
                         let zxRound := add(zx, half)
-                        if lt(zxRound, zx) { revert(0,0) }
+                        if lt(zxRound, zx) { revert(0, 0) }
                         z := div(zxRound, base)
                     }
                 }
